requirements.txt
discord.py==2.3.2
Flask==3.0.3
Flask
Flask